application/serverless_api/cloud_functions/python/login/requirements.txt
bcrypt
mysql-connector-python
google-auth